python/requirements.txt
flask>=2.0.0
libvirt-python
flask-cors
